feat(mobile): implemented UI create page and move previous create to customGoal to reuse for create and customize goal, installed linear gradient, and implement some custom components
