+++ apps/mobile/package.json
@@ -58,7 +58,8 @@
     "react-native-uuid": "^2.0.3",
     "react-native-web": "~0.19.13",
     "react-native-webview": "13.12.5",
-    "zeego": "^2.0.4"
+    "zeego": "^2.0.4",
+    "expo-linear-gradient": "~14.0.2"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
